client module- updated

+++ NJERJIM Guide/Apps/ClientForm.cs
@@ -30,11 +30,6 @@ namespace NJERJIM_Guide
         }
         private void InitializeData()
         {
-            SetDGV();
-        }
-
-        private void SetDGV()
-        {
             var db_helper = new DatabaseHelper();
             var query = "select * from client";
             db_helper.SetDataGridView(clientDataGridView, query);
@@ -56,16 +51,21 @@ namespace NJERJIM_Guide
             {
                 var db_helper = new DatabaseHelper();
                 db_helper.Manipulate($"INSERT INTO {DTClient.Table} ({DTClient.FirstName}, {DTClient.MiddleName}, {DTClient.LastName}, {DTClient.Sex}, {DTClient.ContactNumber}, {DTClient.Addess}) " +
-                    $"VALUES('{firstNameTextBox.Text}', '{middleNameTextBox.Text}', '{lastNameTextBox.Text}', {(int)SelectedSex}, '{contactNumberTextBox.Text}', '{addressTextBox.Text}');");
-                SetDGV();
+                    $"VALUES('{firstNameTextBox.Text}', '{middleNameTextBox.Text}', '{lastNameTextBox.Text}', '{SelectedSex}', '{contactNumberTextBox.Text}', '{addressTextBox.Text}');");
+                InitializeData();
+                clearButton_Click(null,null);
             }
         }
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            var db_helper = new DatabaseHelper();
-            db_helper.Manipulate($"DELETE FROM {DTClient.Table} WHERE {DTClient.Id}={selectedIdLabel.Text};");
-            SetDGV();
+            if (!string.IsNullOrWhiteSpace(selectedIdLabel.Text))
+            {
+                var db_helper = new DatabaseHelper();
+                db_helper.Manipulate($"DELETE FROM {DTClient.Table} WHERE {DTClient.Id}={selectedIdLabel.Text};");
+                InitializeData();
+                clearButton_Click(null, null);
+            }
         }
 
         private void clientDataGridView_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -78,7 +78,7 @@ namespace NJERJIM_Guide
                 firstNameTextBox.Text = selectedRow[1].Value.ToString();
                 middleNameTextBox.Text = selectedRow[2].Value.ToString();
                 lastNameTextBox.Text = selectedRow[3].Value.ToString();
-                if ((Sex)selectedRow[4].Value == Sex.Male) maleRadioButton.Checked = true; else femaleRadioButton.Checked = true;
+                if (selectedRow[4].Value.ToString() == Sex.Male.ToString()) maleRadioButton.Checked = true; else femaleRadioButton.Checked = true;
                 contactNumberTextBox.Text = selectedRow[5].Value.ToString();
                 addressTextBox.Text = selectedRow[6].Value.ToString();
 
@@ -91,6 +91,25 @@ namespace NJERJIM_Guide
         {
             Program.MainMenuForm.Show();
             this.Close();
+        }
+
+        private void addClientTextBox_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.KeyCode == Keys.Enter) addClientButton_Click(null, null);
+        }
+
+        private void clearButton_Click(object sender, EventArgs e)
+        {
+            idLabel.Visible = false;
+            selectedIdLabel.Visible = false;
+            selectedIdLabel.Text = string.Empty;
+            firstNameTextBox.Text = string.Empty;
+            middleNameTextBox.Text = string.Empty;
+            lastNameTextBox.Text = string.Empty;
+            maleRadioButton.Checked = false;
+            femaleRadioButton.Checked = false;
+            contactNumberTextBox.Text = string.Empty;
+            addressTextBox.Text = string.Empty;
         }
     }
 }
